+++ src/main/scala/edu/ie3/simona/agent/EnvironmentRefs.scala
@@ -8,6 +8,7 @@ package edu.ie3.simona.agent
 
 import edu.ie3.simona.event.RuntimeEvent
 import edu.ie3.simona.ontology.messages.{SchedulerMessage, ServiceMessage}
+import edu.ie3.simona.service.ServiceType
 import edu.ie3.simona.service.em.ExtEmDataService
 import edu.ie3.simona.service.ev.ExtEvDataService
 import edu.ie3.simona.service.results.ResultServiceProxy
@@ -42,4 +43,14 @@ final case class EnvironmentRefs(
     loadProfiles: ActorRef[ServiceMessage],
     emDataService: Option[ActorRef[ExtEmDataService.Message]],
     evDataService: Option[ActorRef[ExtEvDataService.Message]],
-)
+) {
+
+  /** Returns references to services by service type.
+    */
+  def serviceMap: Map[ServiceType, ActorRef[ServiceMessage]] =
+    Seq(
+      Some(ServiceType.WeatherService -> weather),
+      Some(ServiceType.LoadProfileService -> loadProfiles),
+      evDataService.map(ref => ServiceType.EvMovementService -> ref),
+    ).flatten.toMap
+}