+++ src/main/scala/edu/ie3/simona/agent/grid/GridAgentData.scala
@@ -19,6 +19,8 @@ import edu.ie3.simona.agent.participant.ParticipantAgent
 import edu.ie3.simona.config.SimonaConfig
 import edu.ie3.simona.event.ResultEvent
 import edu.ie3.simona.model.grid.{GridModel, RefSystem, VoltageLimits}
+import edu.ie3.simona.ontology.messages.ServiceMessage
+import edu.ie3.simona.service.ServiceType
 import edu.ie3.simona.util.ConfigUtil
 import edu.ie3.simona.util.ConfigUtil.{
   EmConfigUtil,
@@ -44,6 +46,8 @@ object GridAgentData {
     * @param environmentRefs
     *   Containing actor references, that are relevant for the environment of
     *   the grid agent.
+    * @param serviceMap
+    *   References to services by service type.
     * @param simonaConfig
     *   Configuration of SIMONA, that is used for.
     * @param resolution
@@ -56,6 +60,7 @@ object GridAgentData {
     */
   final case class GridAgentConstantData(
       environmentRefs: EnvironmentRefs,
+      serviceMap: Map[ServiceType, ActorRef[ServiceMessage]],
       simonaConfig: SimonaConfig,
       resolution: Long,
       simStartTime: ZonedDateTime,